+++ src/main/java/domain/Attendance.java
@@ -1,49 +1,31 @@
 package domain;
 
+import java.time.LocalDate;
+import java.time.LocalTime;
+
 import domain.attendancetime.AttendanceTime;
 import domain.attendancetime.NoShowAttendanceTime;
 import domain.attendancetime.ShowAttendanceTime;
 
-import java.time.DayOfWeek;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
-
 public class Attendance {
     
-    private static final List<DayOfWeek> WEEKENDS = List.of(DayOfWeek.SATURDAY, DayOfWeek.SUNDAY);
-    
-    private final LocalDate attendDate;
+    private final AttendanceDate attendDate;
     private final AttendanceTime attendTime;
     private final AttendanceStatus status;
     
-    private Attendance(final LocalDate attendDate, final AttendanceTime attendTime, final AttendanceStatus status) {
-        validateNotWeekend(attendDate);
-        validateNotHoliday(attendDate);
+    private Attendance(
+            final AttendanceDate attendDate,
+            final AttendanceTime attendTime,
+            final AttendanceStatus status
+    ) {
         this.attendDate = attendDate;
         this.attendTime = attendTime;
         this.status = status;
     }
     
-    private void validateNotWeekend(final LocalDate date) {
-        if (isWeekend(date)) {
-            throw new IllegalArgumentException("주말에는 출석할 수 없습니다.");
-        }
-    }
-    
-    private static boolean isWeekend(final LocalDate date) {
-        return WEEKENDS.contains(date.getDayOfWeek());
-    }
-    
-    private void validateNotHoliday(final LocalDate date) {
-        if (Holiday.isHoliday(date)) {
-            throw new IllegalArgumentException("공휴일에는 출석할 수 없습니다.");
-        }
-    }
-    
     public static Attendance of(final LocalDate attendDate, final LocalTime attendTime) {
         return new Attendance(
-                attendDate,
+                new AttendanceDate(attendDate),
                 new ShowAttendanceTime(attendTime),
                 AttendanceStatus.of(attendDate.getDayOfWeek(), attendTime)
         );
@@ -51,18 +33,14 @@ public class Attendance {
     
     public static Attendance noShow(LocalDate noShowDate) {
         return new Attendance(
-                noShowDate,
+                new AttendanceDate(noShowDate),
                 new NoShowAttendanceTime(),
                 AttendanceStatus.결석
         );
     }
     
-    public static boolean isAttendableDate(LocalDate date) {
-        return !isWeekend(date) && !Holiday.isHoliday(date);
-    }
-    
     public LocalDate getAttendDate() {
-        return attendDate;
+        return attendDate.getAttendDate();
     }
     
     public AttendanceTime getAttendTime() {
